Remove ETag from Tsavorite hot paths & extract ETag RMW to NoInline helpers (Phases 6b, 7)
Phase 6b: Remove dead ETag weight from Tsavorite
- Remove ETag field from RecordMetadata struct (never consumed)
- Remove eTag field from PendingContext
- Remove ~20 pendingContext.eTag assignments from InternalRead/RMW/Upsert/Delete
- Simplify RecordMetadata construction in AllocatorScan, TsavoriteIterator,
  TsavoriteThread, CompletedOutput, Tsavorite.cs

Phase 7: Extract ETag RMW logic to NoInline helpers
- Create RMWMethods.Etags.cs with 9 [NoInlining] helper methods
- Each ETag case in InPlaceUpdaterWorker/InitialUpdater/CopyUpdater/NeedCopyUpdate
  becomes a single-line call, keeping hot-path methods compact for JIT
- Helpers: HandleDelIfGreater*, HandleSetIfMatch*, HandleSetWithEtag*
  for InPlaceUpdate, InitialUpdate, CopyUpdate, and NeedCopyUpdate
