+++ libs/storage/Tsavorite/cs/src/core/Index/Common/RecordMetadata.cs
@@ -15,19 +15,12 @@ namespace Tsavorite.core
         /// </summary>
         public readonly long Address;
 
-        /// <summary>
-        /// The ETag of the record, if any; otherwise <see cref="LogRecord.NoETag"/>.
-        /// </summary>
-        /// <remarks>Included here to be available for multi-key operations.</remarks>
-        public readonly long ETag;
-
-        internal RecordMetadata(long address = kInvalidAddress, long eTag = LogRecord.NoETag)
+        internal RecordMetadata(long address = kInvalidAddress)
         {
             Address = address;
-            ETag = eTag;
         }
 
         /// <inheritdoc/>
-        public override string ToString() => $"addr {AddressString(Address)}, eTag {ETag}";
+        public override string ToString() => $"addr {AddressString(Address)}";
     }
 }
