+++ libs/storage/Tsavorite/cs/src/core/Index/Tsavorite/Implementation/ContinuePending.cs
@@ -52,7 +52,6 @@ namespace Tsavorite.core
                 }
 
                 stackCtx.SetRecordSourceToHashEntry(hlogBase);
-                pendingContext.eTag = pendingContext.diskLogRecord.ETag;
 
                 try
                 {
