+++ libs/storage/Tsavorite/cs/src/core/Index/Tsavorite/Implementation/InternalRMW.cs
@@ -62,7 +62,6 @@ namespace Tsavorite.core
             OperationStackContext<TStoreFunctions, TAllocator> stackCtx = new(keyHash);
             pendingContext.keyHash = keyHash;
             pendingContext.logicalAddress = kInvalidAddress;
-            pendingContext.eTag = LogRecord.NoETag;
 
             if (sessionFunctions.Ctx.phase == Phase.IN_PROGRESS_GROW)
                 SplitBuckets(stackCtx.hei.hash);
@@ -162,7 +161,6 @@ namespace Tsavorite.core
 
                         // status has been set by InPlaceUpdater
                         pendingContext.logicalAddress = stackCtx.recSrc.LogicalAddress;
-                        pendingContext.eTag = srcLogRecord.ETag;
                         goto LatchRelease;
                     }
 
@@ -181,7 +179,6 @@ namespace Tsavorite.core
                         MarkPage(stackCtx.recSrc.LogicalAddress, sessionFunctions.Ctx);
 
                         pendingContext.logicalAddress = stackCtx.recSrc.LogicalAddress;
-                        pendingContext.eTag = srcLogRecord.ETag;
 
                         // Dispose resources and decrement value heap BEFORE setting Tombstone,
                         // so that GetValueHeapMemorySize returns the correct pre-tombstone value.
@@ -207,7 +204,6 @@ namespace Tsavorite.core
                     {
                         // status has been set by InPlaceUpdater, and no modification should have been made to the record.
                         pendingContext.logicalAddress = stackCtx.recSrc.LogicalAddress;
-                        pendingContext.eTag = srcLogRecord.ETag;
                         status = OperationStatusUtils.AdvancedOpCode(OperationStatus.NOTFOUND, StatusCode.WrongType);
                         goto LatchRelease;
                     }
@@ -323,7 +319,6 @@ namespace Tsavorite.core
                         // Success
                         MarkPage(stackCtx.recSrc.LogicalAddress, sessionFunctions.Ctx);
                         pendingContext.logicalAddress = stackCtx.recSrc.LogicalAddress;
-                        pendingContext.eTag = logRecord.ETag;
 
                         // We "IPU'd" because we reused a tombstone, but since the record we have reused did not logically exist, we must also bubble up that the original key was not found (logically).
                         // OperationStatus.NOTFOUND bubbles up success but also indicates that the record was not found in the database.
@@ -702,7 +697,6 @@ namespace Tsavorite.core
             Done:
                 stackCtx.ClearNewRecord();
                 pendingContext.logicalAddress = newLogicalAddress;
-                pendingContext.eTag = newLogRecord.ETag;
                 return status;
             }
